+++ r8te_example.cfg
@@ -1,15 +1,15 @@
 [local]
 # database to be used to keep track of player logged in/out events
-db_name = r8gptDb
-log_file = r8gpt
+db_name = r8teDb
+log_file = r8te
 
-[r8gpt]
+[r8te]
 # Time (in seconds) that bot checks for an updated world save
 scan_time = 75
 # Time (in seconds) between world saves which indicates a likely server reboot event
 # Set this to slightly above your auto save world state rate
 reboot_time = 140
-# Reminder time (in minutes) tells r8gpt how often to send reminders about stuck trains
+# Reminder time (in minutes) tells r8te how often to send reminders about stuck trains
 ai_alert_time = 20
 player_alert_time = 20
 reminder_time = 60
